Update with some new assignments.
I had an absolute path to an image that an instructor pointed out will not work on anyone's computer but mine. I fixed the link to be a relative path so everyone will be able to enjoy the image of Nic Cage. (this was for the Basic_HTML_1.html assignment.