Added pictures to gallery.
Per step 261 of the HTML and CSS course, I am pushing the new changes made to the Gallery section of the index page for my web project. I am doing something else besides the space theme. Mine is going to be a web page where people can apply to be a fighter in the Apex Games (fictional, based on the video game Apex Legends).