One-Page Websitre update.
Pushing per step 383 of the HTML and CSS course. This page was similar in theme to my other Apex Legends themed page so I was able to use most of the css elements from those files in this one too. There were some issues I had to sort out and bugs I had to fix though.